requirements.txt
absl-py==2.3.1
addict==2.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
contourpy==1.3.3
coverage==7.11.0
cycler==0.12.1
einops==0.8.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.9.0
grpcio==1.75.1
hf-xet==1.1.10
huggingface-hub==0.35.3
idna==3.11
imageio==2.37.0
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mmengine==0.10.7
mpmath==1.3.0
networkx==3.5
noise==1.2.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
protobuf==6.32.1
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
rich==14.2.0
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.10.4
tokenizers==0.22.1
torch==2.9.0
torchaudio==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers==4.57.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
yapf==0.43.0
